project/requirements.txt
Django==4.2.0
djangorestframework
django-filter
django-browser-reload
django-celery-beat==2.6.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-ckeditor==6.7.1
django-extensions==3.2.0
django-redis==5.2.0
django-crispy-forms==1.14.0
django-tailwind[reload]
djangorestframework-simplejwt==5.3.1

Pillow
file-validator 
python-dotenv
drf-yasg
transliterate
whitenoise
python-telegram-bot
openai
setuptools
celery
beautifulsoup4
drf-extra-fields==3.7.0
djangorestframework-simplejwt==5.3.1

moviepy
# easyocr
librosa
opencv-python
numpy
scikit-image
imagehash 
instaloader
twilio

#deploy
gunicorn
psycopg2-binary
